+++ src/script/bindings/mod.rs
@@ -17,6 +17,7 @@ pub mod gopher;
 pub mod hash;
 pub mod helpers;
 pub mod imap;
+pub mod ipfs;
 pub mod jwt;
 pub mod netstatus;
 pub mod sample;